+++ demisto_sdk/commands/common/hook_validations/readme.py
@@ -85,7 +85,8 @@ class ReadMeValidator(BaseValidator):
             self.verify_no_default_sections_left(),
             self.verify_readme_is_not_too_short(),
             self.is_context_different_in_yml(),
-            self.verify_demisto_in_readme_content()
+            self.verify_demisto_in_readme_content(),
+            self.verify_template_not_in_readme()
         ])
 
     def mdx_verify(self) -> bool:
@@ -374,6 +375,27 @@ class ReadMeValidator(BaseValidator):
 
         if invalid_lines:
             error_message, error_code = Errors.readme_contains_demisto_word(invalid_lines)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                is_valid = False
+
+        return is_valid
+
+    def verify_template_not_in_readme(self):
+        """
+        Checks if there are the generic sentence '%%FILL HERE%%' in the README content.
+
+        Return:
+            True if '%%FILL HERE%%' does not exist in the README content, and False if it does.
+        """
+        is_valid = True
+        invalid_lines = []
+
+        for line_num, line in enumerate(self.readme_content.split('\n')):
+            if '%%FILL HERE%%' in line:
+                invalid_lines.append(line_num + 1)
+
+        if invalid_lines:
+            error_message, error_code = Errors.template_sentence_in_readme(invalid_lines)
             if self.handle_error(error_message, error_code, file_path=self.file_path):
                 is_valid = False