+++ demisto_sdk/commands/format/update_widget.py
@@ -41,8 +41,8 @@ class WidgetJSONFormat(BaseUpdateJSON):
 
     def format_file(self) -> Tuple[int, int]:
         """Manager function for the widget JSON updater."""
-        format = self.run_format()
-        return format, SKIP_RETURN_CODE
+        format_res = self.run_format()
+        return format_res, SKIP_RETURN_CODE
 
     def set_isPredefined(self):
         """