+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -433,7 +433,8 @@ class TestIntegrationValidation:
         pack_integration_path = join(AZURE_FEED_PACK_PATH, "Integrations/FeedAzure/FeedAzure.yml")
         with ChangeCWD(CONTENT_REPO_EXAMPLE_ROOT):
             runner = CliRunner(mix_stderr=False)
-            result = runner.invoke(main, [VALIDATE_CMD, "-i", pack_integration_path, "--no-conf-json"])
+            result = runner.invoke(main, [VALIDATE_CMD, "-i", pack_integration_path, "--no-conf-json",
+                                          "--allow-skipped"])
 
         assert f"Validating {pack_integration_path} as integration" in result.stdout
         assert "The docker image tag is not the latest numeric tag, please update it" in result.stdout
@@ -455,7 +456,8 @@ class TestIntegrationValidation:
         """
         integration_path = join(TEST_FILES_PATH, 'integration-invalid-no-hidden-params.yml')
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, "-i", integration_path, "--no-conf-json"])
+        result = runner.invoke(main, [VALIDATE_CMD, "-i", integration_path, "--no-conf-json",
+                                      "--allow-skipped"])
         assert result.exit_code == 1
         assert f"Validating {integration_path} as integration" in result.stdout
         assert "can't be hidden. Please remove this field" in result.stdout
@@ -474,7 +476,7 @@ class TestIntegrationValidation:
         """
         integration_path = join(TEST_FILES_PATH, 'integration-valid-no-unallowed-hidden-params.yml')
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, "-i", integration_path, "--no-conf-json"])
+        result = runner.invoke(main, [VALIDATE_CMD, "-i", integration_path, "--no-conf-json", "--allow-skipped"])
         assert f"Validating {integration_path} as integration" in result.stdout
         assert "can't be hidden. Please remove this field" not in result.stdout
         assert result.stderr == ""
@@ -529,7 +531,8 @@ class TestPackValidation:
         mocker.patch.object(BaseValidator, 'check_file_flags', return_value='')
         mocker.patch.object(IntegrationValidator, 'is_there_separators_in_names', return_value=True)
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, "-i", VALID_PACK_PATH, "--no-conf-json"])
+        result = runner.invoke(main, [VALIDATE_CMD, "-i", VALID_PACK_PATH, "--no-conf-json",
+                                      "--allow-skipped"])
         assert f"{VALID_PACK_PATH} unique pack files" in result.stdout
         assert f"Validating pack {VALID_PACK_PATH}" in result.stdout
         assert f"{VALID_PACK_PATH}/Integrations/FeedAzureValid/FeedAzureValid.yml" in result.stdout
@@ -552,7 +555,8 @@ class TestPackValidation:
         mocker.patch.object(ContentEntityValidator, '_load_conf_file', return_value=CONF_JSON_MOCK)
         mocker.patch.object(BaseValidator, 'check_file_flags', return_value='')
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, "-i", AZURE_FEED_PACK_PATH, "--no-conf-json"])
+        result = runner.invoke(main, [VALIDATE_CMD, "-i", AZURE_FEED_PACK_PATH, "--no-conf-json",
+                                      "--allow-skipped"])
 
         assert f'{AZURE_FEED_PACK_PATH}' in result.output
         assert f'{AZURE_FEED_PACK_PATH}/IncidentFields/incidentfield-city.json' in result.output
@@ -1538,7 +1542,8 @@ class TestPlaybookValidation:
         mocker.patch.object(tools, 'is_external_repository', return_value=True)
         mocker.patch.object(PlaybookValidator, 'is_script_id_valid', return_value=True)
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, '-i', VALID_PLAYBOOK_FILE_PATH], catch_exceptions=False)
+        result = runner.invoke(main, [VALIDATE_CMD, '-i', VALID_PLAYBOOK_FILE_PATH, '--allow-skipped',
+                                      '--no-conf-json'], catch_exceptions=False)
         assert f'Validating {VALID_PLAYBOOK_FILE_PATH} as playbook' in result.stdout
         assert 'The files are valid' in result.stdout
         assert result.exit_code == 0
@@ -1557,7 +1562,8 @@ class TestPlaybookValidation:
         mocker.patch.object(tools, 'is_external_repository', return_value=True)
         with ChangeCWD(TEST_FILES_PATH):
             runner = CliRunner(mix_stderr=False)
-            result = runner.invoke(main, [VALIDATE_CMD, '-i', INVALID_PLAYBOOK_FILE_PATH], catch_exceptions=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-i', INVALID_PLAYBOOK_FILE_PATH, '--allow-skipped',
+                                          '--no-conf-json'], catch_exceptions=False)
         assert f'Validating {INVALID_PLAYBOOK_FILE_PATH} as playbook' in result.stdout
         assert 'PB103' in result.stdout
         assert 'The following tasks ids have no previous tasks: {\'5\'}' in result.stdout
@@ -1579,7 +1585,8 @@ class TestPlaybookValidateDeprecated:
         mocker.patch.object(tools, 'is_external_repository', return_value=True)
         mocker.patch.object(PlaybookValidator, 'is_script_id_valid', return_value=True)
         runner = CliRunner(mix_stderr=False)
-        result = runner.invoke(main, [VALIDATE_CMD, '-i', VALID_DEPRECATED_PLAYBOOK_FILE_PATH], catch_exceptions=False)
+        result = runner.invoke(main, [VALIDATE_CMD, '-i', VALID_DEPRECATED_PLAYBOOK_FILE_PATH, '--no-conf-json',
+                                      '--allow-skipped'], catch_exceptions=False)
         assert f'Validating {VALID_DEPRECATED_PLAYBOOK_FILE_PATH} as playbook' in result.stdout
         assert 'The files are valid' in result.stdout
         assert result.exit_code == 0
@@ -1598,8 +1605,8 @@ class TestPlaybookValidateDeprecated:
         mocker.patch.object(tools, 'is_external_repository', return_value=True)
         with ChangeCWD(TEST_FILES_PATH):
             runner = CliRunner(mix_stderr=False)
-            result = runner.invoke(main, [VALIDATE_CMD, '-i', INVALID_DEPRECATED_PLAYBOOK_FILE_PATH],
-                                   catch_exceptions=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-i', INVALID_DEPRECATED_PLAYBOOK_FILE_PATH, '--no-conf-json',
+                                          '--allow-skipped'], catch_exceptions=False)
         assert f'Validating {INVALID_DEPRECATED_PLAYBOOK_FILE_PATH} as playbook' in result.stdout
         assert 'PB104' in result.stdout
         assert 'Deprecated.' in result.stdout
@@ -2502,3 +2509,135 @@ class TestValidationUsingGit:
 
         assert result.exit_code == 1
         assert "You may not be running" in result.stdout  # check error str is in stdout
+
+    def test_validation_using_git_on_specific_file(self, mocker, repo):
+        """
+        Given
+        - A repo with a pack with a modified integration and script.
+
+        When
+        - Running validate using git on it with -i flag aiming at the integration yml path specifically.
+
+        Then
+        - Ensure the integration is validated.
+        - Ensure the script is not validated.
+        """
+        pack = repo.create_pack('FeedAzure')
+        integration = pack.create_integration()
+        integration.create_default_integration()
+        script = pack.create_script()
+        script.create_default_script()
+
+        modified_files = {integration.yml.rel_path, script.yml.rel_path}
+        mocker.patch.object(tools, 'is_external_repository', return_value=False)
+        mocker.patch.object(ValidateManager, 'setup_git_params', return_value=True)
+        mocker.patch.object(ValidateManager, 'setup_prev_ver', return_value='origin/master')
+
+        mocker.patch.object(PackDependencies, 'find_dependencies', return_value={})
+        mocker.patch.object(PackUniqueFilesValidator, 'validate_pack_meta_file', return_value=True)
+        mocker.patch.object(BaseValidator, 'update_checked_flags_by_support_level', return_value=None)
+        mocker.patch.object(ValidateManager, 'get_changed_files_from_git', return_value=(modified_files, set(),
+                                                                                         set(), set()))
+        mocker.patch.object(GitUtil, '__init__', return_value=None)
+        mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        with ChangeCWD(repo.path):
+            runner = CliRunner(mix_stderr=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
+                                          '--skip-pack-release-notes', '-i', integration.yml.rel_path],
+                                   catch_exceptions=False)
+
+        assert 'Running on committed and staged files' in result.stdout
+        assert f'Validating {integration.yml.rel_path}' in result.stdout
+        assert f'Validating {script.yml.rel_path}' not in result.stdout
+
+    def test_validation_using_git_on_specific_file_renamed(self, mocker, repo):
+        """
+        Given
+        - A repo with a pack with a renamed integration and modified script.
+
+        When
+        - Running validate using git on it with -i flag aiming at the integration yml path specifically.
+
+        Then
+        - Ensure the integration is validated.
+        - Ensure the script is not validated.
+        """
+        pack = repo.create_pack('FeedAzure')
+        integration = pack.create_integration()
+        integration.create_default_integration()
+        script = pack.create_script()
+        script.create_default_script()
+
+        modified_files = {(integration.yml.rel_path, integration.yml.rel_path), script.yml.rel_path}
+        mocker.patch.object(tools, 'is_external_repository', return_value=False)
+        mocker.patch.object(ValidateManager, 'setup_git_params', return_value=True)
+        mocker.patch.object(ValidateManager, 'setup_prev_ver', return_value='origin/master')
+
+        mocker.patch.object(PackDependencies, 'find_dependencies', return_value={})
+        mocker.patch.object(PackUniqueFilesValidator, 'validate_pack_meta_file', return_value=True)
+        mocker.patch.object(BaseValidator, 'update_checked_flags_by_support_level', return_value=None)
+        mocker.patch.object(ValidateManager, 'get_changed_files_from_git', return_value=(modified_files, set(),
+                                                                                         set(), set()))
+        mocker.patch.object(GitUtil, '__init__', return_value=None)
+        mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        with ChangeCWD(repo.path):
+            runner = CliRunner(mix_stderr=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
+                                          '--skip-pack-release-notes', '-i', integration.yml.rel_path],
+                                   catch_exceptions=False)
+
+        assert 'Running on committed and staged files' in result.stdout
+        assert f'Validating {integration.yml.rel_path}' in result.stdout
+        assert f'Validating {script.yml.rel_path}' not in result.stdout
+
+    def test_validation_using_git_on_specific_pack(self, mocker, repo):
+        """
+        Given
+        - A repo with two packs.
+
+        When
+        - Running validate using git on it with -i flag aiming at the pack_1 path specifically.
+
+        Then
+        - Ensure the entities in pack 1 are validated
+        - Ensure the entities in pack 2 are not validated.
+        """
+        pack_1 = repo.create_pack('Pack1')
+        integration = pack_1.create_integration()
+        integration.create_default_integration()
+        script = pack_1.create_script()
+        script.create_default_script()
+
+        pack_2 = repo.create_pack('Pack2')
+        integration_2 = pack_2.create_integration()
+        integration_2.create_default_integration()
+        script_2 = pack_2.create_script()
+        script_2.create_default_script()
+
+        modified_files = {(integration.yml.rel_path, integration.yml.rel_path), script.yml.rel_path,
+                          integration_2.yml.rel_path, script_2.yml.rel_path}
+        mocker.patch.object(tools, 'is_external_repository', return_value=False)
+        mocker.patch.object(ValidateManager, 'setup_git_params', return_value=True)
+        mocker.patch.object(ValidateManager, 'setup_prev_ver', return_value='origin/master')
+
+        mocker.patch.object(PackDependencies, 'find_dependencies', return_value={})
+        mocker.patch.object(PackUniqueFilesValidator, 'validate_pack_meta_file', return_value=True)
+        mocker.patch.object(BaseValidator, 'update_checked_flags_by_support_level', return_value=None)
+        mocker.patch.object(ValidateManager, 'get_changed_files_from_git', return_value=(modified_files, set(),
+                                                                                         set(), set()))
+        mocker.patch.object(GitUtil, '__init__', return_value=None)
+        mocker.patch.object(GitUtil, 'get_current_working_branch', return_value='MyBranch')
+
+        with ChangeCWD(repo.path):
+            runner = CliRunner(mix_stderr=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-g', '--no-docker-checks', '--no-conf-json',
+                                          '--skip-pack-release-notes', '-i', pack_1.path],
+                                   catch_exceptions=False)
+
+        assert 'Running on committed and staged files' in result.stdout
+        assert f'Validating {integration.yml.rel_path}' in result.stdout
+        assert f'Validating {script.yml.rel_path}' in result.stdout
+        assert f'Validating {integration_2.yml.rel_path}' not in result.stdout
+        assert f'Validating {script_2.yml.rel_path}' not in result.stdout